Moved code related to clock timer to composable
